🎨 [style] Improve test readability with consistent formatting
- Updates formatting of test assertions and property-based test setups for improved clarity and maintainability.
- Replaces single-line and nested calls with multi-line structures, ensuring better alignment and easier review.
- Does not introduce logic changes; focuses solely on code style and readability for future contributors.

+++ src/test/utils/typeGuards.advanced-fuzzing.test.ts
@@ -215,7 +215,7 @@ describe("Type Guards Advanced Fuzzing Tests", () => {
         fcTest.prop([
             fc.oneof(
                 fc.integer({ min: 1 }),
-                fc.float({ min: 0.0001, noDefaultInfinity: true, noNaN: true }),
+                fc.float({ min: Math.fround(0.0001), noDefaultInfinity: true, noNaN: true }),
                 fc.constant(0),
                 fc.constant(-0),
                 fc.integer({ max: -1 }),
